unify the tag naming for password reset url
Previously, there was [reset_pass_url] and [reset_url]. It was already from before (before the plugin ownership was transferred).

[reset_pass_url] was for setting password for new user
and [reset_url] was for resetting password for existing user

Now, they are unified to [reset_pass_url]. It's chosen over [reset_url] because [reset_pass_url] was more promoted (it's in the template tags metabox in sidebar in the settings page).