chore: remove load_plugin_textdomain() usage
https://developer.wordpress.org/reference/functions/load_plugin_textdomain/#comment-7382

The `load_plugin_textdomain()` call is no longer necessary for plugins hosted on WordPress.org.
Since WP 4.6 translations are auto-loaded from the /languages directory, and since WP 6.7 a “just-in-time” loading system is used.
This change removes the manual call and ensures the plugin uses proper text-domains in __(), _e() etc and places .mo files in the standard languages folder.